% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{regmonkey}
\alias{regmonkey}
\title{RegMonkey: Automated Regression Analysis}
\usage{
regmonkey(
  data,
  dependent_var,
  independent_vars,
  control_vars = NULL,
  mandatory_vars = NULL,
  significance_level = 0.05,
  direction = "both"
)
}
\arguments{
\item{data}{A data frame containing the variables referenced in the formula.}

\item{dependent_var}{A string specifying the name of the dependent variable.}

\item{independent_vars}{A character vector specifying the names of independent variables
to be tested in the model.}

\item{control_vars}{An optional character vector of control variables that may be included
in the model. Default is NULL.}

\item{mandatory_vars}{An optional character vector of variables that must always be included
as controls in every model tested. Default is NULL.}

\item{significance_level}{A numeric value specifying the significance level for the
hypothesis test of the coefficients. Default is 0.05.}

\item{direction}{A string indicating the direction of the effect to test: 'both' for both
positive and negative, 'positive' only for positive, and 'negative' only
for negative effects. Default is 'both'.}
}
\value{
A list of significant combinations. Each element of the list is itself a list
containing the independent variable and the controls under which it was found
significant.
}
\description{
This function performs automated regression analysis by iterating over possible
combinations of control variables, with or without mandatory variables. It assesses
the significance of independent variables' effects on a dependent variable.
}
\examples{
data <- data.frame(y = rnorm(100), x1 = rnorm(100), x2 = rnorm(100), x3 = rnorm(100))
significant_results <- regmonkey(data, "y", c("x1", "x2"), control_vars = c("x3"))

}
