% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_reg}
\alias{summary_reg}
\title{Summarize and Display Regression Results}
\usage{
summary_reg(significant_results, data, coef_map_override = NULL)
}
\arguments{
\item{significant_results}{A list of significant variable combinations returned by \code{regmonkey}. Each element of the list should contain the dependent variable, the independent variable, and the controls under which the significant result was found.}

\item{data}{The data frame containing all the variables referenced in the regression formulas.}

\item{coef_map_override}{An optional named vector to override the default mapping of coefficient names in the output. This can be used to provide more readable or specific names for the variables in the regression output. If not provided, the function will automatically extract variable names from the models.}
}
\value{
This function does not return anything; it directly outputs the regression results formatted by the \code{modelsummary} package to the R console or to an R Markdown document.
}
\description{
This function processes the results from the \code{regmonkey} function, creating and displaying regression models for each significant variable combination identified. It utilizes the \code{modelsummary} package to format and present these results clearly, allowing for an easy evaluation of the significance and impact of various predictors.
}
\examples{
# Assuming `regmonkey` has been run and stored results in `monkey_results`
data <- data.frame(y = rnorm(100), x1 = rnorm(100), x2 = rnorm(100), x3 = rnorm(100))
significant_results <- regmonkey(data, "y", c("x1", "x2"), control_vars = c("x3"))
summary_reg(significant_results, data)
}
